refactor (client): move common elements to chart group component
